+++ CHANGELOG.md
@@ -15,7 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [unreleased]
 
 ### Added
-- Call an other (external) URL/endpoint from a hook
+- Call another (external) URL/endpoint from a hook
 - Native Sentry message support
 - Support for (Inoreader filter) RSS news item notification
 - Support for Freshping monitoring updates